#48 - refactor: 웹소켓 관련 패키지 구조 수정 및 EventListener 코드리팩토링
