+++ generator-api/src/main/java/com/simpaylog/generatorapi/dto/enums/EventType.java
@@ -1,6 +1,7 @@
 package com.simpaylog.generatorapi.dto.enums;
 
 public enum EventType {
+    START,
     COMPLETE,  // 완료
     FAIL, // 실패
     PROGRESS // 진행 중
