+++ generator-api/src/main/java/com/simpaylog/generatorapi/handler/ProgressWebSocketHandler.java
@@ -1,7 +1,7 @@
 package com.simpaylog.generatorapi.handler;
 
 import com.simpaylog.generatorapi.service.SimulationService;
-import com.simpaylog.generatorapi.service.WebSocketSessionManager;
+import com.simpaylog.generatorapi.configuration.websocket.WebsocketSessionManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -16,7 +16,7 @@ import java.time.LocalDate;
 @RequiredArgsConstructor
 public class ProgressWebSocketHandler extends TextWebSocketHandler {
 
-    private final WebSocketSessionManager webSocketSessionManager;
+    private final WebsocketSessionManager webSocketSessionManager;
     private final SimulationService simulationService;
 
     // 클라이언트가 연결되었을 때 실행
@@ -24,7 +24,6 @@ public class ProgressWebSocketHandler extends TextWebSocketHandler {
     public void afterConnectionEstablished(WebSocketSession session) throws Exception {
         log.info("WebSocket session established: {}", session.getId());
         webSocketSessionManager.addSession(session);
-
         simulationService.startSimulation((LocalDate) session.getAttributes().get("durationStart"), (LocalDate) session.getAttributes().get("durationEnd"));
     }
